# openapi.yaml
openapi: 3.1.1
info:
  title: Blog API
  version: 1.0.0
  description: API REST para gestionar autores y posts de un blog

servers:
  - url: http://localhost:3000
    description: Desarrollo local

paths:
  /api/authors:
    get:
      summary: Obtener todos los autores
      tags:
        - Autores
      responses:
        '200':
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Crear un nuevo autor
      tags:
        - Autores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  description: Nombre completo del autor
                email:
                  type: string
                  format: email
                  description: Email del autor (debe ser único)
                bio:
                  type: string
                  description: Biografía del autor
      responses:
        '201':
          description: Autor creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Datos inválidos (nombre y email son requeridos)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: El email ya está registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/authors/{id}:
    get:
      summary: Obtener un autor por ID
      tags:
        - Autores
      parameters:
        - name: id
          in: path
          required: true
          description: ID del autor
          schema:
            type: integer
      responses:
        '200':
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Autor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Actualizar un autor
      tags:
        - Autores
      parameters:
        - name: id
          in: path
          required: true
          description: ID del autor
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nombre completo del autor
                email:
                  type: string
                  format: email
                  description: Email del autor (debe ser único)
                bio:
                  type: string
                  description: Biografía del autor
      responses:
        '200':
          description: Autor actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Autor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: El email ya está registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Eliminar un autor
      description: Elimina el autor y todos sus posts asociados (cascade)
      tags:
        - Autores
      parameters:
        - name: id
          in: path
          required: true
          description: ID del autor
          schema:
            type: integer
      responses:
        '200':
          description: Autor eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Autor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/posts:
    get:
      summary: Obtener posts
      description: Obtiene todos los posts, opcionalmente filtrados por estado de publicación
      tags:
        - Posts
      parameters:
        - name: published
          in: query
          required: false
          description: Filtrar por estado de publicación
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Crear un nuevo post
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
                - author_id
              properties:
                title:
                  type: string
                  description: Título del post
                content:
                  type: string
                  description: Contenido del post
                author_id:
                  type: integer
                  description: ID del autor
                published:
                  type: boolean
                  description: Estado de publicación
                  default: false
      responses:
        '201':
          description: Post creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Datos inválidos (título, contenido y author_id son requeridos)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: El autor especificado no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/posts/{id}:
    get:
      summary: Obtener un post por ID
      tags:
        - Posts
      parameters:
        - name: id
          in: path
          required: true
          description: ID del post
          schema:
            type: integer
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Actualizar un post
      tags:
        - Posts
      parameters:
        - name: id
          in: path
          required: true
          description: ID del post
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Título del post
                content:
                  type: string
                  description: Contenido del post
                published:
                  type: boolean
                  description: Estado de publicación
      responses:
        '200':
          description: Post actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Eliminar un post
      tags:
        - Posts
      parameters:
        - name: id
          in: path
          required: true
          description: ID del post
          schema:
            type: integer
      responses:
        '200':
          description: Post eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/posts/author/{authorId}:
    get:
      summary: Obtener posts de un autor
      tags:
        - Posts
      parameters:
        - name: authorId
          in: path
          required: true
          description: ID del autor
          schema:
            type: integer
      responses:
        '200':
          description: Lista de posts del autor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Author:
      type: object
      properties:
        id:
          type: integer
          description: ID único del autor
        name:
          type: string
          description: Nombre completo del autor
        email:
          type: string
          format: email
          description: Email del autor
        bio:
          type: string
          description: Biografía del autor
        created_at:
          type: string
          format: date-time
          description: Fecha de creación

    Post:
      type: object
      properties:
        id:
          type: integer
          description: ID único del post
        title:
          type: string
          description: Título del post
        content:
          type: string
          description: Contenido del post
        author_id:
          type: integer
          description: ID del autor
        published:
          type: boolean
          description: Estado de publicación
        created_at:
          type: string
          format: date-time
          description: Fecha de creación

    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error

    Message:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de confirmación